♻️ refactor : 코드 가시성 개선 => 기능과 스타일 코드 구분
