--- client/src/data/stacks.js
@@ -1,5 +1,0 @@
-import { sortByASC } from 'hangul-util';
-
-const rowStackList = ['HTML', 'React', 'JavaScript', 'TypeScript', 'CSS', '리액트', 'ReactQuery'];
-
-export const stackList = sortByASC(rowStackList);
