+++ client/src/routes/index.tsx
@@ -1,22 +1,22 @@
 import { Route, Routes as ReactRoutes } from 'react-router-dom';
-import Main from '../pages/Main';
-import Result from '../pages/Result';
+import MainPage from '../pages/MainPage';
+import ResultPage from '../pages/ResultPage';
 import Test from '../pages/Test';
 import MyPage from '../pages/MyPage';
 import Interview from '../pages/Interview';
-import Login from '../pages/Login';
+import LoginPage from '../pages/LoginPage';
 import ReplyInterview from '../pages/ReplyInterview';
 
 const Routes = () => {
   return (
     <ReactRoutes>
-      <Route path="/" element={<Main />} />
+      <Route path="/" element={<MainPage />} />
       <Route path="/test" element={<Test />} />
-      <Route path="/result" element={<Result />} />
+      <Route path="/result" element={<ResultPage />} />
       <Route path="/myPage" element={<MyPage />} />
       <Route path="/interview" element={<Interview />} />
       <Route path="/interview/:id" element={<ReplyInterview />} />
-      <Route path="/login" element={<Login />} />
+      <Route path="/login" element={<LoginPage />} />
     </ReactRoutes>
   );
 };
